Changed mechanism to check served requests
Incoming requests and served requests are both managed using queues.
While queues are exactly the right thing for incoming requests (they
are first come first serve) it is not entirely true for the served ones:
a requests that has been served first is not necessarily the first one
to be queried.

It is not only inefficient looking at the first served request but
it can lead to ENOSPACE if the first request is not queried for a
long time while the other ones continue to write without being able
to issue new requests because still wiating for the first one.

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>