Request queue now allocated dynam and with a new queue wrapper
In order to be ready to allocate more queues when new threads are
spawned, queues need to be allocated dynamically.
This is a first attempt using the request queue (there is only one, so far),
allocated on the heap and with a wrapper queue type containing the
underlying queue and its mutex and condition.

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>