Changed names of default queues
To do any job, the daemon is always supposed to have at least
an incoming queue and an outgoing queue.
A default incoming queue and outgoing queue are allocated since the beginning.

A default incoming queue is surely useful in the "daemon pool" case but
it is not necessary in the standalone case. We will see..

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>